Clean up core/input/keysym.js
KeyTable now only contains the keys needed by core noVNC to represent
KeyboardEvent `key` values.  The full table of keysyms can be found
in FullKeyTable.  We also introduce KeySymMap, which contains mappings
between KeyboardEvent `key` values and keysyms, as well as the new
XF86KeyTable, which contains the full range of XFree86 extra keysyms
(some of which are now included in KeyTable).